requirements.txt
tqdm
lxml
requests
orjson
google-cloud-bigquery
packaging
typing_extensions